update exception response format and improve error handling
