+++ packages/apps/fortune/recording-oracle/src/common/utils/http.ts
@@ -1,0 +1,10 @@
+import { AxiosError } from 'axios';
+
+export function formatAxiosError(error: AxiosError) {
+  return {
+    name: error.name,
+    stack: error.stack,
+    cause: error.cause,
+    message: error.message,
+  };
+}
