+++ packages/apps/fortune/recording-oracle/src/common/utils/webhook.ts
@@ -1,20 +1,11 @@
 import { HttpService } from '@nestjs/axios';
 import { Logger } from '@nestjs/common';
-import { AxiosError } from 'axios';
 import { firstValueFrom } from 'rxjs';
 import { WebhookDto } from '../../modules/webhook/webhook.dto';
 import { HEADER_SIGNATURE_KEY } from '../constants';
 import { CaseConverter } from './case-converter';
 import { signMessage } from './signature';
-
-export function formatAxiosError(error: AxiosError) {
-  return {
-    name: error.name,
-    stack: error.stack,
-    cause: error.cause,
-    message: error.message,
-  };
-}
+import { formatAxiosError } from './http';
 
 export async function sendWebhook(
   httpService: HttpService,
